refactor(#95): 멤버 서비스 분리
